Model Three: Complex and Nasty now implemented however...
There's a bug in the client which means it doesn't shut down properly.
I think it's simply that the agent thread needs to be told to shut down.

I haven't looked into it, but presumably this is caused by the PING/PONG
chatter between client and server.

I've raised this bug as issue #43

Closes 23